backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
openai==1.10.0
anthropic==0.15.0
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==1.1.0
pillow==10.2.0
pytesseract==0.3.10
aiofiles==23.2.1
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
boto3==1.34.0
minio==7.2.0
tiktoken==0.9.0
sentry-sdk[fastapi,celery,redis]==2.19.2